# lib/router.github.js
// /api/v2/oauth/github

let RSAKey = require("../lib/thinx/rsakey.js");
let rsakey = new RSAKey();
let GitHub = require("../lib/thinx/github.js");
let GitHubLink = require("../lib/thinx/github_link.js");
let Sanitka = require("../lib/thinx/sanitka.js"); let sanitka = new Sanitka();

const Globals = require("./thinx/globals");
const prefix = Globals.prefix();
const Database = require("../lib/thinx/database.js");
let db_uri = new Database().uri();
let userlib = require("./thinx/couch")(db_uri).use(prefix + "managed_users"); // lgtm [js/unused-local-variable]

const github_ocfg = Globals.github_ocfg();
const https = require('https');
const sha256 = require("sha256");
const oauthReturn = require("./thinx/oauth_return");
const { readSecret } = require("./thinx/secrets.js");

const app_config = Globals.app_config();

let AuditLog = require("../lib/thinx/audit"); let alog = new AuditLog();

const Util = require("./thinx/util");

//
// OAuth2 for GitHub
//

module.exports = function (app) {

    // SEC-CSRF-05 / D-11: the token link is a cookie-authenticated account mutation.
    // The OAuth initiator and callback GETs below stay unguarded (redirect flow).
    const csrf = require("./middleware/csrf")(app);

    /*
    * OAuth 2 with GitHub
    */

    let user = app.owner;

    // SEC-CFG-02 (D-01, D-02): the client secret comes from
    // /run/secrets/GITHUB_CLIENT_SECRET, then env. Without it no OAuth client
    // is built and the GitHub routes answer 400. The client id is not a
    // secret and stays an env read.
    function buildGithubOAuth() {
        const github_client_secret = readSecret("GITHUB_CLIENT_SECRET");
        if (!github_client_secret) return undefined;
        try {
            let specs = {
                githubClient: process.env.GITHUB_CLIENT_ID,
                githubSecret: github_client_secret,
                baseURL: github_ocfg.base_url, // should be rather gotten from global config!
                loginURI: '/api/oauth/github',
                callbackURI: '/api/oauth/github/callback',
                scope: 'user'
            };
            const client = require('./thinx/oauth-github.js')(specs);
            // CR-01: the one 'error' listener, registered once per client and
            // never per request. It only logs the reason string (never the
            // response, never a secret); each request answers its own errors
            // in the callback passed to githubOAuth.callback().
            client.on('error', (err) => {
                const reason = (err && typeof (err.error) === "string") ? err.error : "unknown";
                console.log(`[debug] [oauth] [github] login error: ${reason}`);
            });
            return client;
        } catch (e) {
            console.log(`[debug] [oauth] [github] github_ocfg init error: ${e}`);
            return undefined;
        }
    }

    let githubOAuth = buildGithubOAuth();

    if (!readSecret("GITHUB_CLIENT_SECRET")) {
        console.log("ℹ️ [info] [router.github] GITHUB_CLIENT_SECRET not set — GitHub OAuth disabled");
    }

    function validateGithubUser(response, token, userWrapper) {

        let owner_id = userWrapper.owner; // must not be nil

        // Check user and make note on user login
        userlib.get(userWrapper.owner, (error, udoc) => {

            // Error case covers creating new user/managing deleted account
            if (error) {

                if (error.toString().indexOf("Error: deleted") !== -1) {
                    console.log("🔨 [debug] [oauth] [check] user document deleted");
                    response.redirect(
                        app_config.public_url + '/error.html?success=failed&title=Sorry&reason=' +
                        encodeURI('Account document deleted.')
                    );
                    return;
                }

                // May exist, but be deleted. Can be cleared using Filtered Replication Handler "del"
                if (typeof (udoc) !== "undefined" && udoc !== null) {
                    if ((typeof (udoc.deleted) !== "undefined") && udoc.deleted === true) {
                        console.log("🔨 [debug] [oauth] [check] user account marked as deleted");
                        response.redirect(
                            app_config.public_url + '/error.html?success=failed&title=Sorry&reason=' +
                            encodeURI('Account deleted.')
                        );
                        return;
                    }
                } else {
                    console.log("No such owner, should create one...");
                }

                // No such owner, create without activation.
                this.user.create(userWrapper, false, response, (/* res, success, status*/) => {

                    console.log("[OID:" + owner_id + "] [NEW_SESSION] [oauth] 2485:");

                    alog.log(owner_id, "OAuth User created. ");

                    app.redis_client.set(token, JSON.stringify(userWrapper));
                    app.redis_client.expire(token, 30);

                    // New account: consent not yet given -> g=false so the console
                    // shows the GDPR consent gate before logging in.
                    const courl = oauthReturn.returnURLFor(response.thx_return_origin, token, false);

                    console.log("Redirecting to login (2) with token", Util.redactToken(token));
                    response.redirect(courl); // for successful login, this must be a response to /oauth/<idp>/callback
                });
                return;
            }

            console.log(`ℹ️ [info] Calling trackUserLogin on GtHub Auth Callback...`);
            user.trackUserLogin(owner_id);

            console.log("validateGithubUser", {
                token: Util.redactToken(token),
                owner: userWrapper.owner
            });
            app.redis_client.set(token, JSON.stringify(userWrapper));
            app.redis_client.expire(token, 3600);

            // gdpr=true means the user has already consented (console then skips
            // the consent page). Keyed on gdpr_consent alone — see needsGDPR in
            // router.google.js for why the old doc.info requirement was dropped.
            let gdpr = false;
            if (typeof (udoc) !== "undefined" && udoc !== null) {
                if (typeof (udoc.gdpr_consent) !== "undefined" && udoc.gdpr_consent === true) {
                    gdpr = true;
                }
            }

            const ourl = oauthReturn.returnURLFor(response.thx_return_origin, token, gdpr); // require GDPR consent
            response.redirect(ourl);
        }); // userlib.get
    }

    function githubLogin(access_token, hdata, res, original_response) {
        let token = "ghat:" + access_token;
        let owner_id, given_name, family_name, email;
    
        if ((typeof (hdata.name) !== "undefined") && hdata.name !== null) {
            let in_name_array = hdata.name.split(" ");
            given_name = in_name_array[0];
            family_name = in_name_array[in_name_array.count - 1];
        } else {
            family_name = hdata.login;
            given_name = hdata.login;
            console.log("🔨 [debug] [github] [token] Warning: no name in GitHub access token response, using login fallback.");
        }
        email = hdata.email || hdata.login;
    
        try {
            owner_id = sha256(prefix + email);
        } catch (e) {
            console.log("☣️ [error] [github] [token] error parsing e-mail: " + e + " email: " + Util.redactEmail(email));
            return res.redirect(app_config.public_url + '/error.html?success=failed&title=Sorry&reason=Missing%20e-mail.');
        }
        validateGithubUser(original_response, token, {
            first_name: given_name,
            last_name: family_name,
            email: email,
            owner: owner_id,
            username: owner_id
        });
    }

    // CR-01: the body of the old per-request githubOAuth.on('token') listener,
    // now called only from the callback of the request that did the exchange.
    // Listeners on the shared client ran for EVERY token, so concurrent logins
    // could redirect one user's browser with another user's session token.
    function handleGithubToken(access_token, original_response) {

        if (!Util.isDefined(access_token)) {
            original_response.status(401).end();
            console.log("[github] oauth_response missing (test or intrusion)");
            return;
        }

        if ((!access_token) || (access_token.indexOf("bad_verification") !== -1)) {
            console.log("🔨 [debug] [github] [token] No token, exiting.");
            original_response.status(401).end();
            return;
        }

        const requestOptions = {
            host: 'api.github.com',
            port: 443,
            path: '/user',
            headers: {
                'User-Agent': 'THiNX', // Application name from GitHub / Settings / Developer Settings 
                'Authorization': 'token ' + access_token, 
                'Accept': 'application/vnd.github+json' 
            } 
        };

        https.get(requestOptions, (res) => {
            let data = '';
            res.on('data', (chunk) => {
                data += chunk;
            });
            res.on('end', () => {
                githubLogin(access_token, JSON.parse(data), res, original_response);
            });
        });
    }

    // Initial page redirecting to OAuth2 provider.
    // Mounted under both /api and /api/v2 — the Vue console builds OAuth URLs from
    // the /api/v2 base (parity with the dual-mounted /login + /logout routes).
    app.get(['/api/oauth/github', '/api/v2/oauth/github'], function (req, res) {
        if (typeof (req.session) !== "undefined") {
            console.log("🔨 [debug] GET /api/oauth/github will destroy old session...");
            req.session.destroy();
        }
        // Remember an allowlisted console origin (e.g. the Vue console) for the callback.
        oauthReturn.rememberReturnOrigin(req, res);
        if (typeof (githubOAuth) !== "undefined") {
            console.log("🔨 [debug] GET /api/oauth/github calling githubOAuth.login");
            githubOAuth.login(req, res);
        } else {
            res.status(400).end();
        }
    });

    // Callback service parsing the authorization token and asking for the access token
    app.get(['/api/oauth/github/callback', '/api/v2/oauth/github/callback'], function (req, res) {
        // Read+clear the remembered console origin here (req is only available on
        // this handler) and stash it on the response, which flows down to
        // validateGithubUser where the post-login redirect is built.
        res.thx_return_origin = oauthReturn.takeReturnOrigin(req, res);

        if (typeof (githubOAuth) === "undefined") {
            console.log("[critical] [githubOAuth] undefined on secure! attempting to fix...");
            githubOAuth = buildGithubOAuth();
        }

        // Still no client: no secret resolved (D-02) or the build failed.
        if (typeof (githubOAuth) === "undefined") {
            res.status(400).end();
            return;
        }

        // The token (or the error) arrives in THIS request's callback only.
        githubOAuth.callback(req, res, (err, access_token) => {
            if (err) {
                // oauth-github.js ends a failed request itself (403/401/502)
                // unless it was answered here. Under ENVIRONMENT=test the specs
                // expect the old 'test-ok' answer.
                if (process.env.ENVIRONMENT == "test" && !res.writableEnded) res.end("test-ok");
                return;
            }
            handleGithubToken(access_token, res);
        });
    });

    // #392 (GH-01 + GH-02): authenticated per-user token link. Validates the
    // caller's GitHub token, stores it on THEIR user document (not a hardcoded
    // test owner), auto-creates an RSA key if they have none, and pushes the
    // public key to GitHub. The token is never echoed back in the response.
    app.post(['/api/github/token', '/api/v2/github/token'], csrf.verifyCsrfToken, (req, res) => {
        if (!Util.validateSession(req)) return res.status(401).end();
        let owner_id = sanitka.owner(req.session.owner);
        if (!owner_id) return Util.responder(res, false, "invalid_owner");

        GitHubLink.link({ GitHub: GitHub, rsakey: rsakey, user: user }, owner_id, req.body.token, (result) => {
            if (result.status && result.status !== 200) {
                return Util.failureResponse(res, result.status, result.response);
            }
            Util.respond(res, { success: result.success, response: result.response });
        });
    });
};