# thinx-api-openapi.yaml
openapi: 3.0.3
info:
  title: THiNX Device Management API
  description: |
    THiNX Device Management API provides functionality for managing IoT devices,
    firmware updates, and related services. The API supports both v1 and v2 endpoints.
  version: 2.0.0
  contact:
    name: THiNX Support
    url: https://thinx.cloud
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://rtm.thinx.cloud/api
    description: Legacy API (v1)
  - url: https://rtm.thinx.cloud/api/v2
    description: Current API (v2)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    
    sessionCookie:
      type: apiKey
      in: cookie
      name: x-thx-core
      description: Session cookie obtained after login
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authentication
      description: API key for device authentication
    
    ownerApiKeyAuth:
      type: object
      description: Owner ID and API key in request body
      properties:
        owner:
          type: string
          description: Owner ID
        api_key:
          type: string
          description: API key
...
  parameters:
    XsrfTokenHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: |
        Anti-CSRF token. Must equal the `XSRF-TOKEN` cookie sent with the same request
        (compared in constant time) and must be bound to the `x-thx-core` session cookie sent
        with it. Obtain it by calling `GET /csrf-token` first and keeping every cookie that
        call sets (`x-thx-core` and `XSRF-TOKEN`) in a cookie jar. The token rotates at login:
        take the new value from the login response's `Set-Cookie`. Browser consoles do this
        automatically. Bearer (JWT) calls, and API-key calls that send no session cookie, are
        exempt and need neither the header nor the cookies. An API key never exempts the login,
        registration or password routes.
      schema:
        type: string
...
  responses:
    CsrfTokenInvalid:
      description: |
        Anti-CSRF check failed: the `X-XSRF-TOKEN` header is missing, does not match the
        `XSRF-TOKEN` cookie, or the token is not bound to the `x-thx-core` session sent with
        it (stale, foreign, or taken from before login). Rejected before the route handler
        runs, so nothing was changed. Re-prime with `GET /csrf-token` (keeping every cookie it
        sets) and retry once.
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              response:
                type: string
                example: csrf_token_invalid
paths:
...
  /user:
    post:
      summary: Create user
      description: |
        Register a new user account. Registration has no machine-client exemption: prime with `GET /csrf-token` first.

        Anti-CSRF (SEC-CSRF-04/05): a cookie-session call must send the session-bound `X-XSRF-TOKEN` header together with the `x-thx-core` and `XSRF-TOKEN` cookies obtained from `GET /csrf-token`; Bearer (JWT) calls are exempt, and so are API-key calls that carry no session cookie, except on the login, registration (`POST /user`) and password routes; a missing or unbound token is refused with 403 `csrf_token_invalid` before the handler runs.
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/XsrfTokenHeader'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: string
                    example: "registered"
        '400':
          description: Bad request
        '403':
          $ref: '#/components/responses/CsrfTokenInvalid'

    delete:
      summary: Delete user
      description: |
        Delete the authenticated user account.

        Anti-CSRF (SEC-CSRF-04/05): a cookie-session call must send the session-bound `X-XSRF-TOKEN` header together with the `x-thx-core` and `XSRF-TOKEN` cookies obtained from `GET /csrf-token`; Bearer (JWT) calls are exempt, and so are API-key calls that carry no session cookie, except on the login, registration (`POST /user`) and password routes; a missing or unbound token is refused with 403 `csrf_token_invalid` before the handler runs.
      tags:
        - Users
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/XsrfTokenHeader'
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: string
                    example: "deleted"
        '401':
          description: Unauthorized
        '403':
          $ref: '#/components/responses/CsrfTokenInvalid'
...
  /oauth/github/callback:
    get:
      summary: GitHub OAuth callback
      description: Handle OAuth callback from GitHub
      tags:
        - OAuth
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        '302':
          description: Redirect after OAuth flow
...
  /github/token:
    post:
      summary: Add GitHub public key
      description: |
        Add a public key to the authenticated user's GitHub account

        Anti-CSRF (SEC-CSRF-04/05): a cookie-session call must send the session-bound `X-XSRF-TOKEN` header together with the `x-thx-core` and `XSRF-TOKEN` cookies obtained from `GET /csrf-token`; Bearer (JWT) calls are exempt, and so are API-key calls that carry no session cookie, except on the login, registration (`POST /user`) and password routes; a missing or unbound token is refused with 403 `csrf_token_invalid` before the handler runs.
      tags:
        - OAuth
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/XsrfTokenHeader'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Key added
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          description: Unauthorized
        '403':
          $ref: '#/components/responses/CsrfTokenInvalid'
...